Refactoring out quadratic function solver for cylinder/capsule ray intersection tests
