+++ src/primitive/cylinder.rs
@@ -3,6 +3,7 @@ use cgmath::prelude::*;
 
 use {Aabb3, Ray3};
 use prelude::*;
+use primitive::util::cylinder_ray_quadratic_solve;
 
 /// Cylinder primitive
 /// Cylinder body is aligned with the Y axis, with local origin in the center of the cylinders.
@@ -80,14 +81,7 @@ where
     S: BaseFloat,
 {
     fn intersects(&self, r: &Ray3<S>) -> bool {
-        let two = S::one() + S::one();
-
-        let a = r.direction.x * r.direction.x + r.direction.z * r.direction.z;
-        let b = two * (r.direction.x * r.origin.x + r.direction.z * r.origin.z);
-        let c = r.origin.x * r.origin.x + r.origin.z * r.origin.z - self.radius * self.radius;
-
-        // check if a is zero -> vector is parallel with y axis
-        if a.is_zero() {
+        if r.direction.x.is_zero() && r.direction.z.is_zero() {
             if r.direction.y.is_zero() {
                 return false;
             }
@@ -95,51 +89,48 @@ where
             return (r.origin.y >= -self.half_height && r.direction.y <= S::zero())
                 || (r.origin.y <= self.half_height && r.direction.y >= S::zero());
         }
-        let four = two + two;
-        let dr = b * b - four * a * c;
-        if dr < S::zero() {
-            return false;
-        }
-        let drsqrt = dr.sqrt();
-        let t1 = (-b + drsqrt) / (two * a);
-        let t2 = (-b - drsqrt) / (two * a);
 
-        if t1 < S::zero() && t2 < S::zero() {
-            return false;
-        }
+        match cylinder_ray_quadratic_solve(r, self.radius) {
+            None => false,
+            Some((t1, t2)) => {
+                if t1 < S::zero() && t2 < S::zero() {
+                    return false;
+                }
 
-        let t = if t1 < S::zero() {
-            t2
-        } else if t2 < S::zero() {
-            t1
-        } else {
-            t1.min(t2)
-        };
+                let t = if t1 < S::zero() {
+                    t2
+                } else if t2 < S::zero() {
+                    t1
+                } else {
+                    t1.min(t2)
+                };
 
-        let pc = r.origin + r.direction * t;
-        if pc.y <= self.half_height && pc.y >= -self.half_height {
-            return true;
-        }
+                let pc = r.origin + r.direction * t;
+                if pc.y <= self.half_height && pc.y >= -self.half_height {
+                    return true;
+                }
 
-        let n = -Vector3::unit_y();
-        let tp = -(self.half_height + r.origin.dot(n)) / r.direction.dot(n);
-        if tp >= S::zero() {
-            let p = r.origin + r.direction * tp;
-            if p.x * p.x + p.z * p.z < self.radius * self.radius {
-                return true;
+                let n = -Vector3::unit_y();
+                let tp = -(self.half_height + r.origin.dot(n)) / r.direction.dot(n);
+                if tp >= S::zero() {
+                    let p = r.origin + r.direction * tp;
+                    if p.x * p.x + p.z * p.z < self.radius * self.radius {
+                        return true;
+                    }
+                }
+
+                let n = Vector3::unit_y();
+                let tb = -(-self.half_height + r.origin.dot(n)) / r.direction.dot(n);
+                if tb >= S::zero() {
+                    let p = r.origin + r.direction * tb;
+                    if p.x * p.x + p.z * p.z < self.radius * self.radius {
+                        return true;
+                    }
+                }
+
+                false
             }
         }
-
-        let n = Vector3::unit_y();
-        let tb = -(-self.half_height + r.origin.dot(n)) / r.direction.dot(n);
-        if tb >= S::zero() {
-            let p = r.origin + r.direction * tb;
-            if p.x * p.x + p.z * p.z < self.radius * self.radius {
-                return true;
-            }
-        }
-
-        false
     }
 }
 
@@ -150,13 +141,7 @@ where
     type Result = Point3<S>;
 
     fn intersection(&self, r: &Ray3<S>) -> Option<Point3<S>> {
-        let two = S::one() + S::one();
-
-        let a = r.direction.x * r.direction.x + r.direction.z * r.direction.z;
-        let b = two * (r.direction.x * r.origin.x + r.direction.z * r.origin.z);
-        let c = r.origin.x * r.origin.x + r.origin.z * r.origin.z - self.radius * self.radius;
-
-        if a.is_zero() {
+        if r.direction.x.is_zero() && r.direction.z.is_zero() {
             if r.direction.y.is_zero() {
                 return None;
             }
@@ -177,50 +162,44 @@ where
             return None;
         }
 
-        let four = two + two;
-        let dr = b * b - four * a * c;
-        if dr < S::zero() {
-            return None;
-        }
-        let drsqrt = dr.sqrt();
-        let t1 = (-b + drsqrt) / (two * a);
-        let t2 = (-b - drsqrt) / (two * a);
+        cylinder_ray_quadratic_solve(r, self.radius).and_then(|(t1, t2)| {
+            if t1 < S::zero() && t2 < S::zero() {
+                None
+            } else {
+                let mut t = if t1 < S::zero() {
+                    t2
+                } else if t2 < S::zero() {
+                    t1
+                } else {
+                    t1.min(t2)
+                };
 
-        if t1 < S::zero() && t2 < S::zero() {
-            return None;
-        }
+                let n = -Vector3::unit_y();
+                let tp = -(self.half_height + r.origin.dot(n)) / r.direction.dot(n);
+                if tp >= S::zero() && tp < t {
+                    let p = r.origin + r.direction * tp;
+                    if p.x * p.x + p.z * p.z < self.radius * self.radius {
+                        t = tp;
+                    }
+                }
 
-        let mut t = if t1 < S::zero() {
-            t2
-        } else if t2 < S::zero() {
-            t1
-        } else {
-            t1.min(t2)
-        };
+                let n = Vector3::unit_y();
+                let tb = -(-self.half_height + r.origin.dot(n)) / r.direction.dot(n);
+                if tb >= S::zero() && tb < t {
+                    let p = r.origin + r.direction * tb;
+                    if p.x * p.x + p.z * p.z < self.radius * self.radius {
+                        t = tb;
+                    }
+                }
 
-        let n = -Vector3::unit_y();
-        let tp = -(self.half_height + r.origin.dot(n)) / r.direction.dot(n);
-        if tp >= S::zero() && tp < t {
-            let p = r.origin + r.direction * tp;
-            if p.x * p.x + p.z * p.z < self.radius * self.radius {
-                t = tp;
+                let pc = r.origin + r.direction * t;
+                if (pc.y > self.half_height) || (pc.y < -self.half_height) {
+                    None
+                } else {
+                    Some(pc)
+                }
             }
-        }
-
-        let n = Vector3::unit_y();
-        let tb = -(-self.half_height + r.origin.dot(n)) / r.direction.dot(n);
-        if tb >= S::zero() && tb < t {
-            let p = r.origin + r.direction * tb;
-            if p.x * p.x + p.z * p.z < self.radius * self.radius {
-                t = tb;
-            }
-        }
-
-        let pc = r.origin + r.direction * t;
-        if (pc.y > self.half_height) || (pc.y < -self.half_height) {
-            return None;
-        }
-        Some(r.origin + r.direction * t)
+        })
     }
 }
